+++ CHANGELOG.md
@@ -1,7 +1,8 @@
 # 2.x branch
 ## 2.2 branch
 ### 2.2.2
-* fixed bug in "album" and posts "categories" widgets.
+* fixed bug in "album" and posts "categories" widgets;
+* you can specify the minimum and maximum fonts for the "popular tags" widget.
 
 ### 2.2.1
 * added index for tags;